Display AD Login Info Not Static String
Rather than displaying the string “Azure AD” for all users with connections to AAD, set the “Provider Login” value to the Displayable ID from the AD info. Fixes the new requirement for a unique value in SonarQube 7.2 and makes it easier to see what AD account is associated with a specific SQ user.